+++ src/Control/Eff/Concurrent/Protocol/EffectfulServer.hs
@@ -36,7 +36,7 @@ import GHC.Stack (HasCallStack)
 --
 -- Instances can by /index types/ for 'Pdu' family directly, or indirectly via the 'ServerPdu' type family.
 --
--- To builder servers serving multiple protocols, use the generic 'Pdu' instances, for which 'EmbedProtocol'
+-- To builder servers serving multiple protocols, use the generic 'Pdu' instances, for which 'Embeds'
 -- instances exist, like 2-,3-,4-, or 5-tuple.
 --
 -- @since 0.24.1